+++ src/main/java/io/github/chris2011/netbeans/minifierbeans/project/ui/options/ProjectOptionsPanel.java
@@ -379,7 +379,6 @@ public final class ProjectOptionsPanel extends JPanel implements ChangeListener 
         enableShortKeyActionConfirmBox = new javax.swing.JCheckBox();
         jPanel3 = new javax.swing.JPanel();
         separatBuild = new javax.swing.JCheckBox();
-        jLabel7 = new javax.swing.JLabel();
         buildJSMinify = new javax.swing.JCheckBox();
         buildCSSMinify = new javax.swing.JCheckBox();
         skipPreExtensionJS = new javax.swing.JCheckBox();
@@ -391,92 +390,88 @@ public final class ProjectOptionsPanel extends JPanel implements ChangeListener 
         skipPreExtensionXML = new javax.swing.JCheckBox();
         buildJSONMinify = new javax.swing.JCheckBox();
         skipPreExtensionJSON = new javax.swing.JCheckBox();
+        jLabel1 = new javax.swing.JLabel();
+        jLabel2 = new javax.swing.JLabel();
 
-        jLayeredPane8.setBorder(javax.swing.BorderFactory.createTitledBorder(javax.swing.BorderFactory.createEtchedBorder(), org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.jLayeredPane8.border.title"), javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Dialog", 1, 12), new java.awt.Color(102, 102, 102))); // NOI18N
+        jLayeredPane8.setBorder(javax.swing.BorderFactory.createTitledBorder(javax.swing.BorderFactory.createEtchedBorder(), org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.jLayeredPane8.border.title"), javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Segoe UI", 0, 12), new java.awt.Color(102, 102, 102))); // NOI18N
 
         addLogToFile.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(addLogToFile, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.addLogToFile.text")); // NOI18N
-        addLogToFile.setOpaque(false);
+        addLogToFile.setContentAreaFilled(false);
         jLayeredPane8.add(addLogToFile);
-        addLogToFile.setBounds(10, 20, 280, 24);
+        addLogToFile.setBounds(10, 20, 280, 20);
 
         enableOutputLogAlert.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(enableOutputLogAlert, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.enableOutputLogAlert.text")); // NOI18N
-        enableOutputLogAlert.setOpaque(false);
+        enableOutputLogAlert.setContentAreaFilled(false);
         jLayeredPane8.add(enableOutputLogAlert);
-        enableOutputLogAlert.setBounds(10, 40, 390, 24);
+        enableOutputLogAlert.setBounds(10, 40, 390, 20);
 
-        jLayeredPane10.setBorder(javax.swing.BorderFactory.createTitledBorder(javax.swing.BorderFactory.createEtchedBorder(), org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.jLayeredPane10.border.title"), javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Dialog", 1, 12), new java.awt.Color(102, 102, 102))); // NOI18N
+        jLayeredPane10.setBorder(javax.swing.BorderFactory.createTitledBorder(javax.swing.BorderFactory.createEtchedBorder(), org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.jLayeredPane10.border.title"), javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Segoe UI", 0, 12), new java.awt.Color(102, 102, 102))); // NOI18N
 
         enableShortKeyAction.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(enableShortKeyAction, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.enableShortKeyAction.text")); // NOI18N
-        enableShortKeyAction.setOpaque(false);
+        enableShortKeyAction.setContentAreaFilled(false);
         jLayeredPane10.add(enableShortKeyAction);
         enableShortKeyAction.setBounds(10, 20, 280, 20);
 
         enableShortKeyActionConfirmBox.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(enableShortKeyActionConfirmBox, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.enableShortKeyActionConfirmBox.text")); // NOI18N
-        enableShortKeyActionConfirmBox.setToolTipText(org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.enableShortKeyActionConfirmBox.toolTipText")); // NOI18N
-        enableShortKeyActionConfirmBox.setOpaque(false);
+        enableShortKeyActionConfirmBox.setContentAreaFilled(false);
         jLayeredPane10.add(enableShortKeyActionConfirmBox);
         enableShortKeyActionConfirmBox.setBounds(10, 40, 350, 20);
 
         separatBuild.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(separatBuild, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.separatBuild.text")); // NOI18N
-        separatBuild.setToolTipText(org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.separatBuild.toolTipText")); // NOI18N
         separatBuild.setActionCommand(org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.separatBuild.actionCommand")); // NOI18N
-        separatBuild.setOpaque(false);
-
-        jLabel7.setFont(new java.awt.Font("Arial", 1, 12)); // NOI18N
-        org.openide.awt.Mnemonics.setLocalizedText(jLabel7, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.jLabel7.text")); // NOI18N
+        separatBuild.setContentAreaFilled(false);
 
         buildJSMinify.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(buildJSMinify, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.buildJSMinify.text")); // NOI18N
-        buildJSMinify.setOpaque(false);
+        buildJSMinify.setContentAreaFilled(false);
 
         buildCSSMinify.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(buildCSSMinify, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.buildCSSMinify.text")); // NOI18N
-        buildCSSMinify.setOpaque(false);
+        buildCSSMinify.setContentAreaFilled(false);
 
         skipPreExtensionJS.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(skipPreExtensionJS, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionJS.text")); // NOI18N
-        skipPreExtensionJS.setToolTipText(org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionJS.toolTipText")); // NOI18N
-        skipPreExtensionJS.setOpaque(false);
+        skipPreExtensionJS.setContentAreaFilled(false);
 
         skipPreExtensionCSS.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(skipPreExtensionCSS, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionCSS.text")); // NOI18N
-        skipPreExtensionCSS.setToolTipText(org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionCSS.toolTipText")); // NOI18N
-        skipPreExtensionCSS.setOpaque(false);
+        skipPreExtensionCSS.setContentAreaFilled(false);
 
         jLabel6.setForeground(new java.awt.Color(102, 102, 255));
         org.openide.awt.Mnemonics.setLocalizedText(jLabel6, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.jLabel6.text")); // NOI18N
 
         buildHTMLMinify.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(buildHTMLMinify, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.buildHTMLMinify.text")); // NOI18N
-        buildHTMLMinify.setOpaque(false);
+        buildHTMLMinify.setContentAreaFilled(false);
 
         skipPreExtensionHTML.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(skipPreExtensionHTML, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionHTML.text")); // NOI18N
-        skipPreExtensionHTML.setToolTipText(org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionHTML.toolTipText")); // NOI18N
-        skipPreExtensionHTML.setOpaque(false);
+        skipPreExtensionHTML.setContentAreaFilled(false);
 
         buildXMLMinify.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(buildXMLMinify, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.buildXMLMinify.text")); // NOI18N
-        buildXMLMinify.setOpaque(false);
+        buildXMLMinify.setContentAreaFilled(false);
 
         skipPreExtensionXML.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(skipPreExtensionXML, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionXML.text")); // NOI18N
-        skipPreExtensionXML.setToolTipText(org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionXML.toolTipText")); // NOI18N
-        skipPreExtensionXML.setOpaque(false);
+        skipPreExtensionXML.setContentAreaFilled(false);
 
         buildJSONMinify.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(buildJSONMinify, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.buildJSONMinify.text")); // NOI18N
-        buildJSONMinify.setOpaque(false);
+        buildJSONMinify.setContentAreaFilled(false);
 
         skipPreExtensionJSON.setBackground(new java.awt.Color(255, 255, 255));
         org.openide.awt.Mnemonics.setLocalizedText(skipPreExtensionJSON, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionJSON.text")); // NOI18N
-        skipPreExtensionJSON.setToolTipText(org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.skipPreExtensionJSON.toolTipText")); // NOI18N
-        skipPreExtensionJSON.setOpaque(false);
+        skipPreExtensionJSON.setContentAreaFilled(false);
+
+        org.openide.awt.Mnemonics.setLocalizedText(jLabel1, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.jLabel1.text")); // NOI18N
+
+        org.openide.awt.Mnemonics.setLocalizedText(jLabel2, org.openide.util.NbBundle.getMessage(ProjectOptionsPanel.class, "ProjectOptionsPanel.jLabel2.text")); // NOI18N
 
         javax.swing.GroupLayout jPanel3Layout = new javax.swing.GroupLayout(jPanel3);
         jPanel3.setLayout(jPanel3Layout);
@@ -490,27 +485,28 @@ public final class ProjectOptionsPanel extends JPanel implements ChangeListener 
                         .addGap(6, 6, 6)
                         .addComponent(jLabel6))
                     .addGroup(jPanel3Layout.createSequentialGroup()
-                        .addComponent(buildJSMinify, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(90, 90, 90)
-                        .addComponent(skipPreExtensionJS, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(jPanel3Layout.createSequentialGroup()
-                        .addComponent(buildCSSMinify, javax.swing.GroupLayout.PREFERRED_SIZE, 140, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(60, 60, 60)
-                        .addComponent(skipPreExtensionCSS, javax.swing.GroupLayout.PREFERRED_SIZE, 300, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(jPanel3Layout.createSequentialGroup()
-                        .addComponent(buildHTMLMinify, javax.swing.GroupLayout.PREFERRED_SIZE, 170, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(30, 30, 30)
-                        .addComponent(skipPreExtensionHTML, javax.swing.GroupLayout.PREFERRED_SIZE, 310, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addComponent(jLabel7)
-                    .addGroup(jPanel3Layout.createSequentialGroup()
                         .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(buildJSONMinify, javax.swing.GroupLayout.PREFERRED_SIZE, 170, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(buildXMLMinify))
-                        .addGap(30, 30, 30)
+                            .addGroup(jPanel3Layout.createSequentialGroup()
+                                .addGap(6, 6, 6)
+                                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                                    .addComponent(buildXMLMinify, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(buildHTMLMinify, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(buildCSSMinify, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(buildJSMinify, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, 110, Short.MAX_VALUE)
+                                    .addComponent(buildJSONMinify, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+                            .addComponent(jLabel1))
+                        .addGap(18, 18, 18)
                         .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(skipPreExtensionXML, javax.swing.GroupLayout.PREFERRED_SIZE, 320, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(skipPreExtensionJSON, javax.swing.GroupLayout.PREFERRED_SIZE, 310, javax.swing.GroupLayout.PREFERRED_SIZE))))
-                .addGap(6, 6, 6))
+                            .addGroup(jPanel3Layout.createSequentialGroup()
+                                .addGap(8, 8, 8)
+                                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                                    .addComponent(skipPreExtensionCSS, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(skipPreExtensionHTML, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(skipPreExtensionXML, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(skipPreExtensionJSON, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 85, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                            .addComponent(jLabel2)
+                            .addComponent(skipPreExtensionJS, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.PREFERRED_SIZE, 87, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         jPanel3Layout.setVerticalGroup(
             jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -519,29 +515,31 @@ public final class ProjectOptionsPanel extends JPanel implements ChangeListener 
                 .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(separatBuild)
                     .addComponent(jLabel6))
-                .addGap(16, 16, 16)
-                .addComponent(jLabel7)
-                .addGap(7, 7, 7)
-                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(buildJSMinify)
-                    .addComponent(skipPreExtensionJS))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 19, Short.MAX_VALUE)
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel1)
+                    .addComponent(jLabel2))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(skipPreExtensionJS)
+                    .addComponent(buildJSMinify))
                 .addGap(6, 6, 6)
-                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(buildCSSMinify)
-                    .addComponent(skipPreExtensionCSS))
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(skipPreExtensionCSS)
+                    .addComponent(buildCSSMinify))
                 .addGap(6, 6, 6)
-                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(buildHTMLMinify)
-                    .addComponent(skipPreExtensionHTML))
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(skipPreExtensionHTML)
+                    .addComponent(buildHTMLMinify))
                 .addGap(6, 6, 6)
-                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(buildXMLMinify)
-                    .addComponent(skipPreExtensionXML))
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(skipPreExtensionXML)
+                    .addComponent(buildXMLMinify))
                 .addGap(6, 6, 6)
-                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(buildJSONMinify)
-                    .addComponent(skipPreExtensionJSON))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(skipPreExtensionJSON)
+                    .addComponent(buildJSONMinify))
+                .addGap(23, 23, 23))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
@@ -552,16 +550,19 @@ public final class ProjectOptionsPanel extends JPanel implements ChangeListener 
                 .addGap(6, 6, 6)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addComponent(jPanel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jLayeredPane8)
-                    .addComponent(jLayeredPane10))
-                .addGap(6, 6, 6))
+                    .addGroup(layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                            .addComponent(jLayeredPane8, javax.swing.GroupLayout.DEFAULT_SIZE, 481, Short.MAX_VALUE)
+                            .addComponent(jLayeredPane10))
+                        .addGap(0, 0, Short.MAX_VALUE)))
+                .addContainerGap())
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGap(6, 6, 6)
+                .addContainerGap()
                 .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(6, 6, 6)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jLayeredPane8, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jLayeredPane10, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
@@ -591,8 +592,9 @@ public final class ProjectOptionsPanel extends JPanel implements ChangeListener 
     public javax.swing.JCheckBox enableOutputLogAlert;
     public javax.swing.JCheckBox enableShortKeyAction;
     public javax.swing.JCheckBox enableShortKeyActionConfirmBox;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel6;
-    private javax.swing.JLabel jLabel7;
     private javax.swing.JLayeredPane jLayeredPane10;
     private javax.swing.JLayeredPane jLayeredPane8;
     private javax.swing.JPanel jPanel3;